fixed all the pylint for spacy related files in pull request #2
added .pylint file for ignore usless pylint errors

moved spacy_model.py classes to their own files, these are: converter.py, dependency.py and token.py for less clutter
