+++ knowledge_graph/__init__.py
@@ -1,0 +1,3 @@
+"""
+word_embeddings module can ...
+"""
