+++ requirements.txt
@@ -3,5 +3,7 @@ gensim~=3.8.3
 pandas~=1.1.3
 networkx~=2.5
 matplotlib~=3.3.2
-spacy
+spacy~=2.3.2
 numpy~=1.19.2
+pytest~=6.1.1
+nltk~=3.5
