+++ word_embedding/dependency.py
@@ -1,0 +1,56 @@
+from enum import Enum, auto
+
+
+class Dependency(Enum):
+    """
+    Dependencies for the word embeddings used in general
+    """
+    root = auto()
+    nsubj = auto()
+    acl = auto()
+    acomp = auto()
+    advcl = auto()
+    advmod = auto()
+    agent = auto()
+    amod = auto()
+    appos = auto()
+    attr = auto()
+    aux = auto()
+    auxpass = auto()
+    case = auto()
+    cc = auto()
+    ccomp = auto()
+    compound = auto()
+    conj = auto()
+    cop = auto()
+    csubj = auto()
+    csubjpass = auto()
+    dative = auto()
+    dep = auto()
+    det = auto()
+    dobj = auto()
+    expl = auto()
+    intj = auto()
+    mark = auto()
+    meta = auto()
+    neg = auto()
+    nn = auto()
+    nounmod = auto()
+    npmod = auto()
+    nsubjpass = auto()
+    nummod = auto()
+    oprd = auto()
+    obj = auto()
+    obl = auto()
+    parataxis = auto()
+    pcomp = auto()
+    pobj = auto()
+    poss = auto()
+    preconj = auto()
+    prep = auto()
+    prt = auto()
+    punct = auto()
+    quantmod = auto()
+    relcl = auto()
+    xcomp = auto()
+    other = auto()
